modified list generation tests to include also set generation
