--- src/test/java/io/javarig/generation/javapredefinedtypes/collection/CollectionGenerationTest.java
@@ -1,15 +1,0 @@
-package io.javarig.generation.javapredefinedtypes.collection;
-
-import io.javarig.RandomInstanceGenerator;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.BeforeEach;
-
-@Slf4j
-public class CollectionGenerationTest {
-    private RandomInstanceGenerator randomInstanceGenerator;
-
-    @BeforeEach
-    public void setUp() {
-        randomInstanceGenerator = new RandomInstanceGenerator();
-    }
-}
